updating tools to load property files
